+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/mode/iconpack/conversion/ui/batch/ui/ValidIconItem.kt
@@ -106,7 +106,7 @@ fun ValidIconItem(
                             selected = false,
                             enabled = icon.iconName.name.isNotEmpty(),
                             onClick = { onPreview(icon) },
-                            iconKey = AllIconsKeys.Actions.Preview,
+                            iconKey = AllIconsKeys.General.IndentDetected,
                             content = {
                                 Text(text = "Preview")
                             },